setup.py: Don't import packages that aren't installed yet.
`setup.py` should never attempt to import packages that are not yet
installed, including packages from the source it is distributing.

In this case, importing `get_version()` from `geraldo.version` was causing
`reportlab` to be imported by side-effect (via `geraldo.__init__` which
imports `geraldo.graphics`, besides other modules).

However there is no guarantee that `reportlab` is already installed when
`setup.py` is run, as it is the job of `setup.py` to specify `reportlab` as a
dependency.